chore: address review comment
