+++ packages/analytics-js-common/src/types/ErrorHandler.ts
@@ -22,4 +22,17 @@ export interface IErrorHandler {
   attachErrorListeners(): void;
 }
 
-export type PreloadedError = Array<any>;
+export type ErrorState = {
+  severity: string;
+  unhandled: boolean;
+  severityReason: { type: string };
+};
+
+export type EventTarget = {
+  localName?: string;
+  dataset?: { loader?: string; isnonnativesdk?: string };
+  src?: string;
+  id?: string;
+};
+
+export type PreloadedError = Array<SDKError | ErrorState>;
